Deal with weird strings from `amdgpu-arch` (#74)
On certain mi325 nodes the reported architecture comes with
a bunch of features separated by a `:`.

This was not expected by the aiter config script.

Hot patch the script in the docker to fix the problem.
The fix will be rolled into the next docker.

Note: we observed this weird strings only on the tensorwave nodes,
so only fixing the 325 configs.